requirements.txt
asgiref==3.6.0
Django==4.1.7
pip==23.0.1
setuptools==67.6.1
sqlparse==0.4.3
tzdata==2023.3
wheel==0.40.0

djangorestframework~=3.14.0
djangorestframework-jwt~=1.11.0